docs: :fire: remove website session (#181)
## Description

The website session doesn't nicely fit into the aim of the course. And
it often had issues and was often rushed because previous sessions took
some time. Instead, I think taking the time with the other sessions is
the better approach and maybe we can eventually include a group project
session for the last day.

Closes #150, closes #159, closes #171, closes #143, closes #144, closes
#145

## Checklist

- [x] Formatted Markdown
- [ ] Rendered website locally

Website has an issue with building, I'll solve in a different PR.